% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlaplace}
\alias{dlaplace}
\title{Laplace density function}
\usage{
dlaplace(x)
}
\arguments{
\item{x}{a real number}
}
\value{
the density at \code{x}
}
\description{
Standard Laplace density function
}
