% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{lla}
\alias{lla}
\title{LLA algorithm applied on regularized least square problems}
\usage{
lla(X, y, pentype, a, lambda, K = 1)
}
\arguments{
\item{X}{design matrix}

\item{y}{responsors}

\item{pentype}{the type of penalties, user can choose from "SCAD", "SICA" or "MCP"}

\item{a}{a parameter in some penalty, all three penalties for chosen possess parameter \code{a}}

\item{lambda}{a regularized parameter which appears in SCAD and MCP, but not in SICA since SICA has no dependence on the regularized parameter.}

\item{K}{the iteration numbers, default is 1}
}
\value{
\code{beta}, namely the estimated parameter vectors for a least square problem
}
\description{
LLA, namely local linear approximation methods, will simplify our regularized least square problems to an adaptive lasso problem. In this setting, user can choose penalty from "SICA", "SCAD" or "MCP".
}
\examples{
\dontrun{
data(list = c("X", "y"))
a = 0.11
lambda = 0.15
pentype = "SICA"

beta_hat <- lla(X, y, pentype, a, lambda)
}
}
