% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{mcp_deri}
\alias{mcp_deri}
\title{First-order derivative of MCP penalty}
\usage{
mcp_deri(t, a, lambda)
}
\arguments{
\item{t}{variable, a real number}

\item{a}{parameter, a real number}

\item{lambda}{parameter, a real number}
}
\value{
a real number which is the first-order derivative of MCP penalty
}
\description{
First-order derivative of MCP penalty
}
