% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ms_cv}
\alias{ms_cv}
\title{Cross validation for model selection with SICA}
\usage{
ms_cv(X, y, pentype, seq_a, seq_lambda)
}
\arguments{
\item{X}{design matrix}

\item{y}{responsors}

\item{pentype}{type of penalty, choose from "SICA", "SCAD" and "MCP"}

\item{seq_a}{a vector of different \code{a}s, where the optimal \code{a} is chosen}

\item{seq_lambda}{a vector of different \code{lambda}s, where the optimal \code{lambda} is chosen}
}
\value{
a list contains optimal \code{a} and optimal \code{lambda}
}
\description{
Cross validation method for model selection with SICAs. Apply this and pick out the optimal parameter \code{a} in SICAs and optimal regularized parameter \code{lambda}.
}
\examples{
\dontrun{
data(list = c("X", "y"))
seq_a <- c(7,9,11)
seq_lambda <- c(0.13,0.15,0.17)
pentype = "SCAD"
# Optimal a and lambda
ms_cv(X, y, pentype, seq_a, seq_lambda)
}
}
