% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ms_data_gen}
\alias{ms_data_gen}
\title{Generate multivariate data for model selection}
\usage{
ms_data_gen(s, n, p, r, beta, sigma)
}
\arguments{
\item{s}{\code{1:s} is the support set of \code{beta}, namely the indexes of true variables}

\item{n}{the number of samples}

\item{p}{the dimensionality of variables}

\item{r}{a correlation parameter in the structured covariance matrix of \code{X}, seen *Model_selection vignette*.}

\item{beta}{parameters on the support set of true variables}

\item{sigma}{a standard deviation parameter in the covariance matrix of \code{epsilon}, seen *Model_selection vignette*.}
}
\value{
a list containing design matrix \code{X} and responsors \code{y}
}
\description{
Generate multivariate data \code{list(X, y)} satisfying the linear regression model, \code{y} = \code{X} \code{beta} + \code{epsilon}, for model selection
}
\examples{
\dontrun{
test1_data <- ms_data_gen(s,10000,p,r,beta,sigma1)
test1_X <- test1_data[[1]]
test1_y <- test1_data[[2]]
}
}
