% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwM}
\alias{rwM}
\title{Random walk Metropolis sampler}
\usage{
rwM(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation for Gaussian increment}

\item{x0}{initial point of random walk}

\item{N}{the numer of wanted samples}
}
\value{
a list containing the samples and the number of rejected points
}
\description{
generate random samples from standard laplace distribution using random walk Metropolis sampler
}
