% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{sica_deri}
\alias{sica_deri}
\title{First-order derivative of SICA penalty}
\usage{
sica_deri(t, a)
}
\arguments{
\item{t}{variable, a real number}

\item{a}{parameter, a real number}
}
\value{
a real number which is the first-order derivative of SICA penalty
}
\description{
First-order derivative of SICA penalty
}
