\name{StatComp20026-package}
\alias{StatComp-package}
\alias{StatComp}
\docType{package}
\title{
  A package designed in the course **Statistical Computing**
}
\description{
  *StatComp20026* is designed in the course **Statistical Computing** which is advised by Prof. Hong Zhang and assisted by M.S. Anran Liu. The package is mainly divided into two parts. The first part is of several useful functions helping to solve model selection problem with penaties like SCAD, SICA and MCP. The second part contains all the homework in this course. 
}
\details{
  Details will be shown in vigenettes.
}
\author{
Tingnan Gong

Maintainer: Tingnan Gong <twgtn@mail.ustc.edu.cn>
}
\references{
  See vignettes.
}
\keyword{ package }

